refactor(test): add test clone suite for all scales
